Fix some issues with input event passing
This MR fixes several issues with InputThing and how it passes events to
the compositor.

I removed the listening of events from
QVirtualKeyboardInputEngine::virtualKeyClicked. We already get the key
events in InputThing through keyPressEvent/keyReleaseEvent, and so we
ended up with duplicate events. In keyPressEvent/keyReleaseEvent, the
call of `keysym` was incorrect, as it passed the Qt key enum, not the
xkbcommon keysym value to wayland. With this fixed, I dropped
`virtualKeyClicked` handling (where keysym was implemented properly, but
not for all keys). This fixes the backspace and shift keys not working
properly.

I added an IGNORED_KEYS list for key events to ignore, which fixes the
"special keys" button on the vkbd entering a strange character.

I fixed cursor positioning in `inputMethodEvent`. The
input_method_v1 protocol states that the preedit_cursor call must be
before preedit_string
(https://wayland.app/protocols/input-method-unstable-v1#zwp_input_method_context_v1:request:preedit_cursor). The order was flipped in the code, causing cursor
positioning to be off and delayed by a character.

I uncommented the line in `inputMethodQuery` that implements the "no
predictive text" flag. Text fields with that flag set now immediately
get committed characters without preediting (ex. qmlkonsole)

I added a missing null check in `keysym` that caused crashes if the
input method context was not initialized.

I also added differentiation for key press/release handling between
textual and non-textual buttons. text-input-v3 clients seem to not
handle when keys are entered through keysym, so we need to properly
commit text characters. This fixes text input not working in GTK and
Chromium.