Use QWindow::setMask to allow popups to extend out of window boundary
Fixes: https://invent.kde.org/plasma/plasma-keyboard/-/issues/9

This use [setMask](https://doc.qt.io/qt-6/qwindow.html#setMask) to set a
part of the window to be interactive. KWin will only reserve this area
for the input method panel, so anything rendered above will be overlaid
over the application receiving input.

![image](/uploads/b557e821e5bd7ebd09caacb4f3263778/image.png){width=552 height=988}

I think we need some better contrast for the keyboard theme popovers, but that's a different issue to address