Allow configuration the keyboard as a floating window
Only override QT_WAYLAND_INPUT_PANEL_TOPLEVEL when it was not set.
If it was set to 0 explicitly, enable the floating window mode.
In that mode, other windows are overlayed and the keyboard window
only uses half the available screen width.

The motivation for this is to use the keyboard in widescreen
configuration, where a full-width keyboard can be far too large.