Add support for non-full width operation
This adds a new config option `panelFillScreenWidth=true/false` to plasmakeyboardrc. When set to false, the panel doesn't fill the entire screen width, instead it is centered on the bottom and the width is set to a percentage against the height (constant is set to 3 * height).

I made necessary improvements to the window mask to support having non
full-width panels.