Add support for non-full width operation, and improve floating mode
This extends https://invent.kde.org/plasma/plasma-keyboard/-/merge_requests/31

We still retain two modes, "toplevel" and "floating" modes.
- "Toplevel" - KWin handles keyboard placement always being at the
  bottom of the screen;
- "Floating" - Keyboard placement can be customized by the user, KWin
  does not reserve space

I added a new config option`panelScreenWidthPct` to allow
customization of how wide the panel is as a percentage of the screen
width. By default we have 1.0 (100%), but some users (like @mwolff) may
want it to be 50%. In the future, we can allow users to resize it by
dragging on the edges.

For the floating mode, I wrapped the keyboard panel with a "window"
widget. I added a drag handle at the top to allow users to move the
panel around.

I made necessary improvements to the window mask to support having non
full-width panels.