Create qml plugin
Create a QML plugin to put our own types that are used in styles and
layouts.

This moves some of the shared components in the Breeze style out to
the QML plugin. In the future, if we extend key behavior or make new
keys, we can put them here too to be imported for use in layouts (this
will be necessary for
https://invent.kde.org/plasma/plasma-keyboard/-/issues/26,
https://invent.kde.org/plasma/plasma-keyboard/-/issues/14).